Make rllib optional dependency
