+++ requirements.txt
@@ -1,6 +1,5 @@
 numpy
 torch
 pyglet<=1.5.27
-ray[rllib]<=2.1,>=1.13
 gym==0.23.1
 wandb
